pyproject.toml
[tool.poetry]
name = "multisource-rag-agent"
version = "2.0.0"
description = "Hybrid RAG chatbot with multi-source data retrieval across databases, cloud storage, and document stores"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# Web framework
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
streamlit = "^1.31.0"
httpx = "^0.26.0"

# LLM
openai = "^1.12.0"

# Data validation & config
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

# Structured databases
psycopg2-binary = "^2.9.9"
google-cloud-bigquery = "^3.17.0"
pyodbc = "^5.1.0"
pymysql = "^1.1.0"

# Azure services
azure-storage-blob = "^12.19.0"
azure-cosmos = "^4.5.0"

# Google Cloud Storage
google-cloud-storage = "^2.14.0"

# Document processing
pypdf = "^4.0.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.2"

# Embeddings & vector store
tiktoken = "^0.6.0"
chromadb = "^0.4.22"

# Utilities
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
